+++ src/collective/task/interfaces.py
@@ -5,7 +5,12 @@ from zope import schema
 from zope.interface import Interface
 from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 from plone.app.dexterity import PloneMessageFactory as _PMF
+from plone.autoform import directives
 from plone.supermodel import model
+
+from dexterity.localrolesfield.field import LocalRolesField
+
+from collective.task import _
 
 
 class ICollectiveTaskLayer(IDefaultBrowserLayer):
@@ -20,6 +25,20 @@ class ITaskContent(model.Schema):
         required=True
     )
 
+    parents_assigned_groups = LocalRolesField(
+        title=_(u"Parents assigned groups"),
+        required=False,
+        value_type=schema.Choice(vocabulary=u'collective.task.AssignedGroups')
+    )
+    #directives.mode(parents_assigned_groups='hidden')
+
+    parents_enquirers = LocalRolesField(
+        title=_(u"Parents enquirers"),
+        required=False,
+        value_type=schema.Choice(vocabulary=u'collective.task.Enquirer')
+    )
+    #directives.mode(parents_enquirers='hidden')
+
 
 class ITaskMethods(Interface):
 
@@ -28,3 +47,7 @@ class ITaskMethods(Interface):
             Get the object containing the highest ITask object
             or the highest ITask object if task is True
         """
+
+
+class ITaskContentMethods(Interface):
+    """ Adapter description """