+++ src/collective/task/subscribers.py
@@ -1,5 +1,8 @@
 # -*- coding: utf-8 -*-
 
+from collective.task.adapters import TaskContentAdapter
+from collective.task.interfaces import ITaskContent
+
 
 def afterTransitionITaskSubscriber(obj, event):
     """ After transition subscribers on ITask"""
@@ -14,3 +17,23 @@ def afterTransitionITaskSubscriber(obj, event):
     if event.transition and event.transition.id == 'back_in_to_assign':
         # Remove auto_to_do_flag on task.
         obj.auto_to_do_flag = False
+
+
+def taskContent_created(task, event):
+    """
+        Update parents localrole fields.
+        Moving a tree will trigger events from children to parents !
+    """
+    if event.oldParent is None or event.oldName is None:
+        status = 'creation'
+    elif event.oldParent != event.newParent:
+        status = 'move'
+    elif event.oldName != event.newName:
+        status = 'rename'
+        return
+    print "MOVED %s on %s" % (status, task.absolute_url_path())
+    adapted = TaskContentAdapter(task)
+    if status == 'move':
+        adapted.set_all_parents_value('parents_assigned_groups', 'get_parents_assigned_groups')
+
+    adapted.set_parents_value('parents_assigned_groups', adapted.get_parents_assigned_groups(), modified=False)